Revert .gitattributes
This commit reverts 9804135d300e3801be066c156650d45978c18b51 and
51020009ac983a51d564cbbebab91c72fdaa6065.

These files are not a binary, they are text, a shell scripts. Making it binary
in .gitattributes forces git to hide its diff in `git log --patch` (as it
qualifies it as binary, and it don't show changes for binary files).

Windows users should see
https://help.github.com/articles/dealing-with-line-endings/